N.O.E.L. Parser unpacking implementation.
The parser is now able to unpack the virtual card entry from a file into variables.
This is done with the bitstrings and bit segments which assumes that each character is an ASCII character of 7-bits long with an unused 8th bit.
This is where the values `ID:80/bitstring` come from, 10 char wide, each char 7 + 1 bit long = 80.